quirks/mediaDevices.html

<!doctype html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
    test(function() {
      assert_equals(navigator.mediaDevices.constructor, MediaDevices);
    }, 'navigator.mediaDevices.constructor');
</script>
